cma_space_req_struct.real_addr is now an off_t
This hopefully signals that real_addr should be passed to the mmap

+++ cma_malloc.h
@@ -10,7 +10,7 @@
 
 struct cma_space_request_struct {
     size_t size;                /* Size in bytes */
-    uintptr_t real_addr;        /* The physical address */
+    off_t real_addr;            /* The physical address */
     uintptr_t kern_addr;        /* The virtual address in kernel space */
 };
